@@ -6,33 +6,30 @@
     "scripts": {
         "dev": "vitepress dev --host 0.0.0.0",
         "build": "vitepress build",
-        "serve": "vitepress serve"
+        "preview": "vitepress preview"
     },
     "keywords": [],
     "author": "",
     "license": "ISC",
+    "type": "module",
     "devDependencies": {
-        "@types/node": "^18.11.8",
-        "@typescript-eslint/eslint-plugin": "^5.41.0",
-        "@typescript-eslint/parser": "^5.41.0",
-        "@typescript-eslint/types": "^5.41.0",
-        "@vue/compiler-sfc": "^3.2.41",
-        "eslint": "8.11.0",
-        "eslint-config-prettier": "^8.5.0",
+        "@types/node": "^20.1.5",
+        "@typescript-eslint/eslint-plugin": "^5.59.6",
+        "@typescript-eslint/parser": "^5.59.6",
+        "@typescript-eslint/types": "^5.59.6",
+        "@vue/compiler-sfc": "^3.3.2",
+        "eslint": "^8.40.0",
+        "eslint-config-prettier": "^8.8.0",
         "eslint-plugin-prettier": "^4.2.1",
-        "eslint-plugin-vue": "^9.7.0",
-        "flexsearch": "^0.7.31",
-        "fs-extra": "^10.1.0",
-        "globby": "^11.1.0",
+        "eslint-plugin-vue": "^9.13.0",
+        "fs-extra": "^11.1.1",
+        "globby": "^13.1.4",
         "gray-matter": "^4.0.3",
-        "markdown-it": "^13.0.1",
-        "markdown-it-katex": "^2.0.3",
-        "markdown-it-latex": "^0.2.0",
-        "markdown-it-mathjax3": "^4.3.2",
-        "markdown-it-sub": "^1.0.0",
-        "prettier": "^2.7.1",
-        "typescript": "^4.8.4",
-        "vitepress": "^1.0.0-alpha.26",
-        "vitepress-plugin-search": "^1.0.4-alpha.15"
+        "prettier": "^2.8.8",
+        "typescript": "^5.0.4",
+        "vitepress": "1.0.0-alpha.75"
+    },
+    "dependencies": {
+        "markdown-it-katex": "^2.0.3"
     }
 }
