@@ -1,22 +1,20 @@
 <template>
-    <div style="padding-top: 10px">
-        <div v-for="yearList in data">
-            <div class="year">
-                {{ yearList[0].frontMatter.date.split('-')[0] }}
-            </div>
-            <a :href="withBase(article.regularPath)" v-for="(article, index) in yearList" :key="index" class="article">
-                <div class="titleDiv">
-                    <div class="title-o"></div>
-                    {{ article.frontMatter.title }}
-                </div>
-                <div class="date">{{ article.frontMatter.date.slice(5) }}</div>
-            </a>
+    <div v-for="yearList in data">
+        <div class="year">
+            {{ yearList[0].frontMatter.date.split('-')[0] }}
         </div>
+        <a :href="withBase(article.regularPath)" v-for="(article, index) in yearList" :key="index" class="posts">
+            <div class="post-container">
+                <div class="post-dot"></div>
+                {{ article.frontMatter.title }}
+            </div>
+            <div class="date">{{ article.frontMatter.date.slice(5) }}</div>
+        </a>
     </div>
 </template>
 
 <script lang="ts" setup>
-import { useData,withBase } from 'vitepress'
+import { useData, withBase } from 'vitepress'
 import { computed } from 'vue'
 import { useYearSort } from '../functions'
 
@@ -26,9 +24,9 @@ const data = computed(() => useYearSort(theme.value.posts))
 
 <style scoped>
 .year {
-    padding: 16px 0 8px 0;
-    font-size: 1.2rem;
-    font-weight: 600;
+    padding: 14px 0 8px 0;
+    font-size: 1.25rem;
+    font-weight: 500;
     font-family: var(--date-font-family);
 }
 </style>
