@@ -4,7 +4,7 @@ import NewLayout from './components/NewLayout.vue'
 import Archives from './components/Archives.vue'
 import Tags from './components/Tags.vue'
 import Page from './components/Page.vue'
-// import Comment from './components/Comment.vue'
+import Comment from './components/Comment.vue'
 
 import './custom.css'
 
@@ -16,6 +16,6 @@ export default {
         app.component('Tags', Tags)
         app.component('Archives', Archives)
         app.component('Page', Page)
-        // app.component('Comment', Comment)
+        app.component('Comment', Comment)
     }
 }
